Moved `gendecoders` to the `ipbus` command group See #69
