+++ src/ipbb/cmds/common.py
@@ -85,107 +85,3 @@ def addrtab(env, aDest):
         ),
         fg='green',
     )
-
-
-# ------------------------------------------------------------------------------
-def gendecoders(env, aCheckUpToDate):
-
-    lDecodersDir = 'decoders'
-
-    with DirSentry(env.currentproj.path):
-        sh.rm('-rf', lDecodersDir)
-        # Gather address tables
-        addrtab(env, aDest=lDecodersDir)
-
-    lGenScript = 'gen_ipbus_addr_decode'
-    lGenToolPath = '/opt/cactus/bin/uhal/tools'
-    lGenToolLibPath = '/opt/cactus/lib'
-
-    if not which(lGenScript):
-
-        lPaths = os.environ['PATH'].split() if os.environ['PATH'] else []
-        if lGenToolPath not in lPaths:
-            lPaths[0:0] = [lGenToolPath]
-            os.environ['PATH'] = ':'.join(lPaths)
-
-        if not which(lGenScript):
-            raise click.ClickException("'{0}' script not found.".format(lGenScript))
-
-    lLibPaths = (
-        os.environ['LD_LIBRARY_PATH'].split() if 'LD_LIBRARY_PATH' in os.environ else []
-    )
-    if lGenToolLibPath not in lLibPaths:
-        lLibPaths[0:0] = [lGenToolLibPath]
-        os.environ['LD_LIBRARY_PATH'] = ':'.join(lLibPaths)
-
-    secho("Using " + which(lGenScript), fg='green')
-
-    # ------------------------------------------------------------------------------
-
-    lUpdatedDecoders = []
-    lGen = sh.Command(lGenScript)
-    with DirSentry(join(env.currentproj.path, lDecodersDir)):
-        for lAddr in env.depParser.commands['addrtab']:
-            echo("Processing " + style(basename(lAddr.filepath), fg='blue'))
-            # Interested in top-level address tables only
-            if not lAddr.toplevel:
-                secho(
-                    "{} is not a top-level address table. Decoder will not be generated.".format(
-                        lAddr.filepath
-                    ),
-                    fg='cyan',
-                )
-                continue
-
-            # Generate a new decoder file
-            lGen(basename(lAddr.filepath), _out=sys.stdout, _err_to_out=True)
-            lDecoder = 'ipbus_decode_{0}.vhd'.format(
-                splitext(basename(lAddr.filepath))[0]
-            )
-            lTarget = env.pathMaker.getPath(
-                lAddr.package, lAddr.component, 'src', lDecoder
-            )
-
-            diff = sh.colordiff if which('colordiff') else sh.diff
-
-            # Has anything changed?
-            try:
-                diff('-u', '-I', '^-- START automatically', lTarget, lDecoder)
-            except sh.ErrorReturnCode as e:
-                print(e.stdout.decode())
-
-                lUpdatedDecoders.append((lDecoder, lTarget))
-
-        # ------------------------------------------------------------------------------
-        # If no difference between old and newly generated decoders, quit here.
-        if not lUpdatedDecoders:
-            secho(
-                "\n{}: All ipbus decoders are up-to-date.\n".format(
-                    env.currentproj.name
-                ),
-                fg='green',
-            )
-            return
-
-        # ------------------------------------------------------------------------------
-        echo(
-            'The following decoders have changed and must be updated:\n'
-            + '\n'.join(map(lambda s: '* ' + style(s[0], fg='blue'), lUpdatedDecoders))
-            + '\n'
-        )
-        if aCheckUpToDate:
-            raise SystemExit(-1)
-
-        confirm('Do you want to continue?', abort=True)
-        for lDecoder, lTarget in lUpdatedDecoders:
-            print(sh.cp('-av', lDecoder, lTarget))
-
-        secho(
-            "\n\n{}: {} decoders updated.\n".format(
-                env.currentproj.name, len(lUpdatedDecoders)
-            ),
-            fg='green',
-        )
-
-
-# ------------------------------------------------------------------------------
